requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altgraph
annotated-types==0.7.0
anyio==4.4.0
ascii-magic==2.3.0
astor==0.8.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.13.3
blag==2.2.1
blinker==1.9.0
build==1.2.2.post1
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cssselect2==0.8.0
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
duckdb==1.2.2
et-xmlfile==1.1.0
exceptiongroup==1.2.2
Faker==19.13.0
feedgenerator==2.1.0
feedparser==6.0.11
Flask==3.1.0
fonttools==4.56.0
frozendict==2.4.6
frozenlist==1.4.1
future 
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==8.4.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.9.0
kiwisolver==1.4.7
macholib
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.4
mdurl==0.1.2
multidict==6.0.5
multitasking==0.0.11
ngrok==1.4.0
numpy==2.0.1
ollama==0.3.1
openai==1.77.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==1.5.3
pandasai==2.4.2
peewee==3.17.9
pillow==10.4.0
pip-tools==7.4.1
platformdirs==4.3.7
psutil==7.0.0
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.18.0
pyparsing==3.2.1
pyperclip==1.9.0
pyproject_hooks==1.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
scipy==1.13.1
selenium==4.31.0
sgmllib3k==1.0.0
six
slack==0.0.2
slack_bolt==1.23.0
slack_sdk==3.35.0
slackclient==2.9.4
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.40
sqlglot==25.34.1
sqlglotrs==0.3.0
termcolor==2.4.0
text-unidecode==1.3
tinycss2==1.4.0
tomli==2.2.1
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.1
undetected-chromedriver==3.5.5
urllib3==2.3.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.9.4
yfinance==0.2.55
zipp==3.20.1